Refactor file and folder structure to be more readable
