--- frontend/src/components/Root.js
@@ -1,15 +1,0 @@
-import React from "react";
-import { HashRouter } from "react-router-dom";
-import { Provider } from "react-redux";
-
-import App from "./App";
-
-const Root = React.memo(({ store }) => (
-  <Provider store={store}>
-    <HashRouter>
-      <App />
-    </HashRouter>
-  </Provider>
-));
-
-export default Root;
