+++ frontend/src/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 
 import configureStore from "./store/store";
-import Root from "./components/Root";
+import App from "./App";
 
 document.addEventListener("DOMContentLoaded", () => {
   let store = configureStore();
@@ -24,5 +24,5 @@ document.addEventListener("DOMContentLoaded", () => {
   window.dispatch = store.dispatch;
 
   const root = document.getElementById("root");
-  ReactDOM.render(<Root store={store} />, root);
+  ReactDOM.render(<App store={store} />, root);
 });
